task query追加

+++ annofabcli/task/delete_tasks.py
@@ -1,6 +1,6 @@
 import argparse
 import logging
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Optional
 
 import requests
 from annofabapi.models import ProjectMemberRole, TaskStatus
@@ -8,6 +8,10 @@ from annofabapi.models import ProjectMemberRole, TaskStatus
 import annofabcli
 from annofabcli import AnnofabApiFacade
 from annofabcli.common.cli import AbstractCommandLineInterface, ArgumentParser, build_annofabapi_resource_and_login
+
+from annofabcli.common.facade import TaskQuery, match_task_with_task_query
+
+from annofabapi.dataclass.task import Task
 
 logger = logging.getLogger(__name__)
 
@@ -31,7 +35,7 @@ class DeleteTask(AbstractCommandLineInterface):
         annotation_list = self.service.wrapper.get_all_annotation_list(project_id, query_params=query_params)
         return annotation_list
 
-    def delete_task(self, project_id: str, task_id: str, force: bool = False):
+    def delete_task(self, project_id: str, task_id: str, force: bool = False, task_query: Optional[TaskQuery] = None,):
         """
         タスクを削除します。
 
@@ -70,13 +74,17 @@ class DeleteTask(AbstractCommandLineInterface):
                 logger.info(f"アノテーションが付与されているため（{len(annotation_list)}個）、タスク'{task_id}'を削除できません。")
                 return False
 
+        if not match_task_with_task_query(Task.from_dict(task), task_query):
+            logger.debug(f"task_id={task_id}: TaskQueryの条件にマッチしないため、スキップします。")
+            return False
+
         if not self.confirm_delete_task(task_id):
             return False
 
         self.service.api.delete_task(project_id, task_id)
         return True
 
-    def delete_task_list(self, project_id: str, task_id_list: List[str], force: bool = False):
+    def delete_task_list(self, project_id: str, task_id_list: List[str], force: bool = False, task_query: Optional[TaskQuery] = None):
         """
         複数のタスクを削除する。
         """
@@ -118,6 +126,8 @@ def parse_args(parser: argparse.ArgumentParser):
     argument_parser.add_project_id()
     argument_parser.add_task_id()
     parser.add_argument("--force", action="store_true", help="アノテーションが付与されているタスクも強制的に削除します。")
+    argument_parser.add_task_query()
+
     parser.set_defaults(subcommand_func=main)
